+++ vkconfig/configurator.cpp
@@ -52,9 +52,7 @@ Configurator::Configurator(const std::string &VULKAN_SDK)
 Configurator::~Configurator() {
     configurations.SaveAllConfigurations(layers.available_layers);
 
-    if (environment.GetMode() != LAYERS_MODE_BY_CONFIGURATOR_PERSISTENT) {
-        SurrenderConfiguration(environment);
-    }
+    SurrenderConfiguration(environment);
 }
 
 bool Configurator::Init() {